refactor(models): relocate SwcInternalBehavior classes to spec Package modules (Rule 0007 A)
