+++ src/armodel/models/M2/AUTOSARTemplates/SWComponentTemplate/SwcInternalBehavior/DataElements/InstanceRefsUsage.py
@@ -5,123 +5,8 @@ in software component internal behavior templates.
 
 from typing import List, Optional
 
-from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.ArObject import ARObject
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.PrimitiveTypes import RefType
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.AbstractStructure import AtpInstanceRef
-
-
-class ArVariableInImplementationDataInstanceRef(ARObject):
-    """
-    A reference to an AUTOSAR variable in the context of an implementation
-    data type instance.
-    """
-
-    # ArVariableInImplementationDataInstanceRef method parity checklist:
-    # [ ] __init__                     [x] impl  [ ] docstring  [ ] test
-    # [ ] getContextDataPrototypeRefs  [x] impl  [x] docstring  [ ] test
-    # [ ] setContextDataPrototypeRefs  [x] impl  [x] docstring  [ ] test
-    # [ ] getPortPrototypeRef          [x] impl  [x] docstring  [ ] test
-    # [ ] setPortPrototypeRef          [x] impl  [x] docstring  [ ] test
-    # [ ] getRootVariableDataPrototypeRef [x] impl  [x] docstring  [ ] test
-    # [ ] setRootVariableDataPrototypeRef [x] impl  [x] docstring  [ ] test
-    # [ ] getTargetDataPrototypeRef    [x] impl  [x] docstring  [ ] test
-    # [ ] setTargetDataPrototypeRef    [x] impl  [x] docstring  [ ] test
-
-    def __init__(self):
-        super().__init__()
-
-        self.contextDataPrototypeRefs: List[RefType] = []
-        self.portPrototypeRef: RefType = None
-        self.rootVariableDataPrototypeRef: RefType = None
-        self.targetDataPrototypeRef: RefType = None
-
-    def getContextDataPrototypeRefs(self):
-        """
-        Gets the list of context data prototype references.
-
-        Returns:
-            List[RefType]: The list of context data prototype references
-        """
-        return self.contextDataPrototypeRefs
-
-    def setContextDataPrototypeRefs(self, value):
-        """
-        Sets the list of context data prototype references.
-
-        Args:
-            value: The list of context data prototype references to set
-
-        Returns:
-            self for method chaining
-        """
-        self.contextDataPrototypeRefs = value
-        return self
-
-    def getPortPrototypeRef(self):
-        """
-        Gets the port prototype reference.
-
-        Returns:
-            RefType: The port prototype reference
-        """
-        return self.portPrototypeRef
-
-    def setPortPrototypeRef(self, value):
-        """
-        Sets the port prototype reference.
-
-        Args:
-            value: The port prototype reference to set
-
-        Returns:
-            self for method chaining
-        """
-        self.portPrototypeRef = value
-        return self
-
-    def getRootVariableDataPrototypeRef(self):
-        """
-        Gets the root variable data prototype reference.
-
-        Returns:
-            RefType: The root variable data prototype reference
-        """
-        return self.rootVariableDataPrototypeRef
-
-    def setRootVariableDataPrototypeRef(self, value):
-        """
-        Sets the root variable data prototype reference.
-
-        Args:
-            value: The root variable data prototype reference to set
-
-        Returns:
-            self for method chaining
-        """
-        self.rootVariableDataPrototypeRef = value
-        return self
-
-    def getTargetDataPrototypeRef(self):
-        """
-        Gets the target data prototype reference.
-
-        Returns:
-            RefType: The target data prototype reference
-        """
-        return self.targetDataPrototypeRef
-
-    def setTargetDataPrototypeRef(self, value):
-        """
-        Sets the target data prototype reference.
-
-        Args:
-            value: The target data prototype reference to set
-
-        Returns:
-            self for method chaining
-        """
-        self.targetDataPrototypeRef = value
-        return self
 
 
 class VariableInAtomicSWCTypeInstanceRef(AtpInstanceRef):
@@ -411,67 +296,3 @@ class ParameterInAtomicSWCTypeInstanceRef(AtpInstanceRef):
             The target data prototype reference, or None if not set
         """
         return self.targetDataPrototypeRef
-
-
-class AutosarParameterRef(ARObject):
-    """
-    A reference to an AUTOSAR parameter.
-    """
-
-    # AutosarParameterRef method parity checklist:
-    # [ ] __init__                     [x] impl  [ ] docstring  [ ] test
-    # [ ] getAutosarParameterIRef      [x] impl  [x] docstring  [ ] test
-    # [ ] setAutosarParameterIRef      [x] impl  [x] docstring  [ ] test
-    # [ ] getLocalParameterRef         [x] impl  [x] docstring  [ ] test
-    # [ ] setLocalParameterRef         [x] impl  [x] docstring  [ ] test
-
-    def __init__(self):
-        super().__init__()
-
-        self.autosarParameterIRef: "ParameterInAtomicSWCTypeInstanceRef" = None
-        self.localParameterRef: RefType = None
-
-    def getAutosarParameterIRef(self):
-        """
-        Gets the AUTOSAR parameter instance reference.
-
-        Returns:
-            ParameterInAtomicSWCTypeInstanceRef: The AUTOSAR parameter instance
-                reference
-        """
-        return self.autosarParameterIRef
-
-    def setAutosarParameterIRef(self, value):
-        """
-        Sets the AUTOSAR parameter instance reference.
-
-        Args:
-            value: The AUTOSAR parameter instance reference to set
-
-        Returns:
-            self for method chaining
-        """
-        self.autosarParameterIRef = value
-        return self
-
-    def getLocalParameterRef(self):
-        """
-        Gets the local parameter reference.
-
-        Returns:
-            RefType: The local parameter reference
-        """
-        return self.localParameterRef
-
-    def setLocalParameterRef(self, value):
-        """
-        Sets the local parameter reference.
-
-        Args:
-            value: The local parameter reference to set
-
-        Returns:
-            self for method chaining
-        """
-        self.localParameterRef = value
-        return self
