+++ src/armodel/models/M2/AUTOSARTemplates/SWComponentTemplate/SwcInternalBehavior/Trigger.py
@@ -10,7 +10,7 @@ from typing import TYPE_CHECKING, Optional
 from armodel.models.M2.MSR.DataDictionary.DataDefProperties import SwImplPolicyEnum
 from armodel.models.M2.AUTOSARTemplates.SWComponentTemplate.SwcInternalBehavior.AccessCount import AbstractAccessPoint
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.ArObject import ARObject
-from armodel.models.M2.AUTOSARTemplates.SWComponentTemplate.RPTScenario import IdentCaption
+from armodel.models.M2.AUTOSARTemplates.SWComponentTemplate.RPTScenario import ExternalTriggeringPointIdent
 
 if TYPE_CHECKING:
     from armodel.models.M2.AUTOSARTemplates.SWComponentTemplate.Components.InstanceRefs import PTriggerInAtomicSwcTypeInstanceRef
@@ -54,18 +54,6 @@ class InternalTriggeringPoint(AbstractAccessPoint):
         if value is not None:
             self.swImplPolicy = value
         return self
-
-
-class ExternalTriggeringPointIdent(IdentCaption):
-    """
-    Identification of an external triggering point.
-    """
-
-    # ExternalTriggeringPointIdent method parity checklist:
-    # [ ] __init__                     [x] impl  [ ] docstring  [ ] test
-
-    def __init__(self, parent, short_name):
-        super().__init__(parent, short_name)
 
 
 class ExternalTriggeringPoint(ARObject):
