+++ tests/test_armodel/models/M2/AUTOSARTemplates/SWComponentTemplate/SwcInternalBehavior/DataElements/test_InstanceRefsUsage.py
@@ -3,9 +3,11 @@ This module contains comprehensive tests for the InstanceRefsUsage module in SWC
 Tests cover all classes and methods in the InstanceRefsUsage.py file to achieve 100% test coverage.
 """
 
-from armodel.models.M2.AUTOSARTemplates.SWComponentTemplate.SwcInternalBehavior.InstanceRefsUsage import (
+from armodel.models.M2.AUTOSARTemplates.SWComponentTemplate.SwcInternalBehavior.DataElements import (
     ArVariableInImplementationDataInstanceRef,
     AutosarParameterRef,
+)
+from armodel.models.M2.AUTOSARTemplates.SWComponentTemplate.SwcInternalBehavior.DataElements.InstanceRefsUsage import (
     ParameterInAtomicSWCTypeInstanceRef,
     VariableInAtomicSWCTypeInstanceRef,
 )
@@ -154,7 +156,7 @@ class TestAutosarParameterRef:
         assert param_ref.localParameterRef is None
 
         # Test autosarParameterIRef methods
-        from armodel.models.M2.AUTOSARTemplates.SWComponentTemplate.SwcInternalBehavior.InstanceRefsUsage import ParameterInAtomicSWCTypeInstanceRef
+        from armodel.models.M2.AUTOSARTemplates.SWComponentTemplate.SwcInternalBehavior.DataElements.InstanceRefsUsage import ParameterInAtomicSWCTypeInstanceRef
 
         iref = ParameterInAtomicSWCTypeInstanceRef()
         param_ref.setAutosarParameterIRef(iref)
