+++ tests/test_armodel/models/M2/AUTOSARTemplates/SWComponentTemplate/test_EndToEndProtection.py
@@ -4,14 +4,14 @@ Tests cover all classes and methods in the EndToEndProtection.py file to achieve
 """
 
 from armodel.models.M2.AUTOSARTemplates.AutosarTopLevelStructure import AUTOSAR
-from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.PrimitiveTypes import NameToken, PositiveInteger, RefType
+from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.PrimitiveTypes import NameToken, PositiveInteger
 from armodel.models.M2.AUTOSARTemplates.SWComponentTemplate.EndToEndProtection import (
     EndToEndDescription,
     EndToEndProtection,
-    EndToEndProtectionISignalIPdu,
     EndToEndProtectionSet,
     EndToEndProtectionVariablePrototype,
 )
+from armodel.models.M2.AUTOSARTemplates.SystemTemplate.EndToEndProtection import EndToEndProtectionISignalIPdu
 from armodel.models.M2.AUTOSARTemplates.SystemTemplate.InstanceRefs import VariableDataPrototypeInSystemInstanceRef
 
 
@@ -98,35 +98,6 @@ class TestEndToEndProtectionVariablePrototype:
         iref = VariableDataPrototypeInSystemInstanceRef()
         prototype.addReceiverIref(iref)
         assert iref in prototype.getReceiverIrefs()
-
-
-class TestEndToEndProtectionISignalIPdu:
-    """Test class for EndToEndProtectionISignalIPdu class."""
-
-    def test_end_to_end_protection_i_signal_i_pdu_initialization(self):
-        """Test EndToEndProtectionISignalIPdu initialization and methods."""
-        pdu = EndToEndProtectionISignalIPdu()
-        assert pdu.dataOffset is None
-        assert pdu.iSignalGroupRef is None
-        assert pdu.iSignalIPduRef is None
-
-        # Test setters and getters
-        from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.PrimitiveTypes import Integer
-
-        offset = Integer()
-        offset.setValue(10)
-        pdu.setDataOffset(offset)
-        assert pdu.getDataOffset() == offset
-
-        group_ref = RefType()
-        group_ref.setValue("/Test/ISignalGroup")
-        pdu.setISignalGroupRef(group_ref)
-        assert pdu.getISignalGroupRef() == group_ref
-
-        pdu_ref = RefType()
-        pdu_ref.setValue("/Test/ISignalIPdu")
-        pdu.setISignalIPduRef(pdu_ref)
-        assert pdu.getISignalIPduRef() == pdu_ref
 
 
 class TestEndToEndProtection:
