+++ tests/test_armodel/writer/test_writer_sw_data_def_props.py
@@ -21,7 +21,7 @@ from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.
     RefType,
     TRefType,
 )
-from armodel.models.M2.AUTOSARTemplates.SWComponentTemplate.SwcInternalBehavior.AutosarVariableRef import AutosarVariableRef
+from armodel.models.M2.AUTOSARTemplates.SWComponentTemplate.SwcInternalBehavior.DataElements import AutosarVariableRef
 from armodel.models.M2.MSR.DataDictionary.DataDefProperties import (
     CompuGenericMath,
     DisplayPresentationEnum,
